+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- SqlScript
+  - Added integration test configuration that creates script files and executes
+    the resource in a single configuration using `DependsOn`.
+
 ## [17.5.1] - 2026-02-05
 
 ### Added